Add dashboard view
This commit creates a new dashboard view by implementing the routing
and the component to show.
It creates a new dashboard component with ng generate component
command. It add the proper styles and create the method to grab 4
heroes from the list by using the heroes service.
Finally it adds the dashboard component to the routing module and
creates a link to navigate through both pages.