Enable HTTP service and mock data server
We use the HttpClient module from Angular by including it in the app
module. This would be useful to comunicate with the server.

To avoid creating the server yet, we simulate the responses and the
DB with the in-memory Web API library. We install it with
`npm install angular-in-memory-web-api --save` and generate the
proper service with `ng generate service ...`.